+++ CHANGES.md
@@ -1,2 +1,3 @@
 Release with following new features and fixes:
+- [WebInterface-OnBoot](https://github.com/rM-self-serve/webinterface-onboot) is currently not supported by reMarkable software version 3.16 or higher
 - Build against Avalonia 11.2.3